check for file existence

+++ siflib/io/ska_wrapper.py
@@ -18,15 +18,17 @@ def run_ska(pdb1: str,
     subdir.mkdir(exist_ok=True)
     outfile = subdir / f"{pdb1}-vs-{pdb2}"
     cmd = f"{skabin} {pdb1_path} {pdb2_path}"
-    with outfile.open("w") as of:
-        subprocess.run(cmd, shell=True, env=env, stdout=of, stderr=STDOUT)
+    if not outfile.exists():
+        with outfile.open("w") as of:
+            subprocess.run(cmd, shell=True, env=env, stdout=of, stderr=STDOUT)
 
     subdir = out_dir / pdb2[1:3]
     subdir.mkdir(exist_ok=True)
     outfile = subdir / f"{pdb2}-vs-{pdb1}"
     cmd = f"{skabin} {pdb1_path} {pdb2_path}"
-    with outfile.open("w") as of:
-        subprocess.run(cmd, shell=True, env=env, stdout=of, stderr=STDOUT)
+    if not outfile.exists():
+        with outfile.open("w") as of:
+            subprocess.run(cmd, shell=True, env=env, stdout=of, stderr=STDOUT)
 
 
 def run(query_info: Path,
